move transaction state into own folder
